20.2: Broken intermediate state
Just keeping this for the record.
Thought it was an elegant idea to keep a map from intervals to
accumulated shifts.
Turns out it's too hard to get the details right, the distinction
between index 0 and 6 is too blurry.
If it's so hard it surely can't really be all that elegant?