Solve day 22 part 2 in 102.9 ms
Hardest part was getting all the various coordinate transformations
correct for all the various nets of a cube. Needed tests to get the
details right.

To enable testing parts of the code I refactored the relevant parts into
a separate library and a test suite. Because of a bug in an interaction
between Cabal and GHC I had to move this out of the day's directory.

The colon in the directory name is interpreted by GHC as a colon
separating a list of directories passed to the `-i` flag. Tracking the
issue in haskell/cabal#8687.