Solve day 11 part 2 in 2.4 s
First attempt was to memoize to the actual list of output stones. This
takes too much memory to store and involves more iterating over lists to
count the number of stones.

My original idea was to memoize to the number of output stones instead
and that does work, just required a bit more thinking because you can't
use the memoized values to recurse on, but that turns out not to be
necessary.

Speeds up part 1 to 35.3 ms.