AoC: Relax list to Foldable in aStar
A* doesn't seem feasible on part 2. Current formulation captures both
parts 1 and 2 with only the neighbor function differing.

A different way of looking at the problem is as specifying linear
systems of equations, which turns it into minimizing integer linear
programs. This is where sbv and Z3 come into the equation.